requirements.txt
langchain==0.3.26
langchain-community==0.3.26
faiss-cpu==1.11.0
sentence-transformers==4.1.0
openai==1.90.0
streamlit==1.46.0
arxiv==2.2.0
pymupdf==1.26.1
python-dotenv==1.1.0
neo4j==5.28.1